feat: define password domain
